% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(actual, predicted, threshold = 0.5)
}
\arguments{
\item{actual}{A vector containing the observed class of an observation}

\item{predicted}{A vector containing the predicted class or probability
of class membership of an observation}

\item{threshold}{Optional parameter which is the probability threshold
at which an observation is classified as a member of class 1.}
}
\description{
This function computes the accuracy of predictions (proportion of correct predictions).
}
