% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{brier_score}
\alias{brier_score}
\title{Brier Score}
\usage{
brier_score(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed class of an observation}

\item{predicted}{A vector containing the predicted probabilities}
}
\description{
This function computes the brier score of predictions.
}
