% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conf_mat}
\alias{conf_mat}
\title{Confusion Matrix}
\usage{
conf_mat(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed class of a point}

\item{predicted}{A vector containing the predicted probabilities}
}
\description{
This function displays a confusion matrix of predictions.
}

