% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f1_score}
\alias{f1_score}
\title{F 1 score}
\usage{
f1_score(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed class of an observation}

\item{predicted}{A vector containing the predicted class or probability
of class membership at an observation}
}
\description{
This function computes the F 1 score for predictions.
}
