% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fbeta_score}
\alias{fbeta_score}
\title{F beta score}
\usage{
fbeta_score(actual, predicted, beta = 1)
}
\arguments{
\item{actual}{A vector containing the observed class of an observation}

\item{predicted}{A vector containing the predicted class or probability
of class membership at an observation}

\item{beta}{Specifies which F score you want to calculate (defaults to 1)}
}
\description{
This function computes the F beta score for predictions.
}
