% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming_dist}
\alias{hamming_dist}
\title{Hamming Distance}
\usage{
hamming_dist(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed class of an observation}

\item{predicted}{A vector containing the predicted class.}
}
\description{
This function computes the hamming distance of predictions.
}
