% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_abs_error}
\alias{mean_abs_error}
\title{Mean absolute error}
\usage{
mean_abs_error(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed value of a point}

\item{predicted}{A vector containing the predicted value at a point}
}
\description{
This function will calculate the mean absolute error between two vectors.
}
