% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{precision}
\alias{precision}
\title{Precision}
\usage{
precision(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed class of a point}

\item{predicted}{A vector containing the predicted class}
}
\description{
This function computes the precision of predictions.
}

