% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r2_score}
\alias{r2_score}
\title{R2 score}
\usage{
r2_score(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed value of a point}

\item{predicted}{A vector containing the predicted value at a point}
}
\description{
This function will calculate the r2 score between two vectors.
}
