% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recall}
\alias{recall}
\title{Recall}
\usage{
recall(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed class of an observation}

\item{predicted}{A vector containing the predicted class}
}
\description{
This function computes the recall of predictions.
}
