% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed value of a point}

\item{predicted}{A vector containing the predicted value at a point}
}
\description{
This function will calculate the root mean squared error between two vectors.
}
