% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squared_error}
\alias{squared_error}
\title{Squared error}
\usage{
squared_error(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed value of a point}

\item{predicted}{A vector containing the predicted value at a point}
}
\description{
This function will calculate the element-wise squared error between two vectors
}

