% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tru_neg_rate}
\alias{tru_neg_rate}
\title{True Negative Rate}
\usage{
tru_neg_rate(actual, predicted)
}
\arguments{
\item{actual}{A vector containing the observed class of an observation}

\item{predicted}{A vector containing the predicted class}
}
\description{
This function computes the true negative rate of predictions.
}
