% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_error.R
\name{xgb_error}
\alias{xgb_error}
\alias{xgb_error.character}
\alias{xgb_error.default}
\alias{xgb_error.function}
\title{Xgboost Error}
\usage{
xgb_error(err_func)

\method{xgb_error}{default}(err_func)

\method{xgb_error}{function}(err_func)

\method{xgb_error}{character}(err_func)
}
\arguments{
\item{err_func}{Either a function a character vector of length one containinng the name
of the function. The error function must return a single value.}
}
\value{
A function which can be used within xgboost.
}
\description{
A wrapper function to allow your error metrics to be used as evaluation metrics
for xgboost. Note that this does not mean it is the objective function,
it is only used as an evaluation (checks the error as the algorithm progresses)
}

