% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schema.R
\name{ExampleData}
\alias{ExampleData}
\title{Example AIRR data}
\format{
\code{extdata/rearrangement-example.tsv.gz}: Rearrangement TSV file.
}
\description{
Example data files compliant with the the AIRR Data Representation standards.
}
\examples{
# Get path to the rearrangement-example file
file <- system.file("extdata", "rearrangement-example.tsv.gz", package="airr")

# Load data file
df <- read_rearrangement(file)

}
